nrps_pks, nrps_pks_domains: move domain drawing to the detection module where they're found
