+++ antismash/detection/nrps_pks_domains/templates/details.html
@@ -1,6 +1,6 @@
 <div class="details">
   <h3>Detailed annotation</h3>
-    {% if cluster.get_domain_details() %}
+    {% if has_domain_details %}
       <div class="details-svg" id="{{cluster.anchor_id}}-details-svg"></div>
     {% endif %}
 </div>
